requirements.txt
python=3.8
pandas
google-api-python-client
oauth2client
snowflake-snowpark-python